start to fix batch tests
